got a 404 on that URL
when I clicked on the "You can contribute to this manual by making changes to the source markdown file." link in the manual, I got a 404. This URL is definitely broken. Now this is not a markdown file, so I am pretty sure this is not the most correct way of fixing this. But I could not find the md.